Keep window helpers in their own files
